langgraph/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["orchestrator"]

[project]
name = "langgraph-orchestrator"
version = "0.1.0"
description = "LangGraph orchestrator with OTel tracing"
requires-python = ">=3.11"
dependencies = [
    # LangChain v1+ (latest versions)
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=0.2.68",
    "langsmith>=0.2.0",

    # OpenAI
    "openai>=1.83.0",

    # OpenTelemetry - LangGraph uses native OTEL via environment variables
    "opentelemetry-api>=1.34.0",
    "opentelemetry-sdk>=1.34.0",
    "opentelemetry-exporter-otlp>=1.34.0",
    "openinference-instrumentation-langchain>=0.1.0",

    # MCP for CrewAI tools
    "mcp>=1.16.0",
    "langchain-mcp-adapters>=0.1.0",

    # HTTP client for future distributed tracing
    "httpx>=0.28.0",

    # Utilities
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
